refactor(uat): separate uats wrt commands
