+++ uat/test_uat_init.py
@@ -1,25 +1,6 @@
 import subprocess as sp
 from pathlib import Path
 
-import pytest
-
-
-@pytest.fixture
-def change_test_dir(tmpdir, monkeypatch):
-    monkeypatch.chdir(tmpdir)
-    return tmpdir
-
-
-def test_list_template():
-    check_list_template = sp.run(["gdk", "component", "list", "--template"], check=True, stdout=sp.PIPE)
-    assert "HelloWorld-python" in check_list_template.stdout.decode()
-    assert "HelloWorld-java" in check_list_template.stdout.decode()
-
-
-def test_list_repository():
-    check_list_template = sp.run(["gdk", "component", "list", "--repository"], check=True, stdout=sp.PIPE)
-    assert "aws-greengrass-labs-database-influxdb" in check_list_template.stdout.decode()
-
 
 def test_init_template_non_empty_dir():
     check_init_template = sp.run(["gdk", "component", "init", "-t", "HelloWorld", "-l", "python"], stdout=sp.PIPE)
